lib/scrapers/telegram/requirements.txt
telethon==1.34.0
python-dotenv==1.0.1
supabase==2.3.4
asyncio
